Move connection handler interfaces to proxy/common/connhandler
